% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helper_funcs.R
\docType{data}
\name{theme_project}
\alias{theme_project}
\title{Associate a theme with this project}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
theme_project
}
\description{
Associate a theme with this project
}
\keyword{datasets}
